Moved DatabaseFactory.kt to com.bpavuk.di
